backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.1
pydantic-settings>=2.2.1
groq>=0.9.0
PyPDF2>=3.0.1
pdfplumber>=0.11.0
python-docx>=1.1.2
networkx>=3.3
python-dotenv>=1.0.1
httpx>=0.27.0